Insert Products data

+++ database/migrations/1611122027130_product_schema.js
@@ -9,7 +9,7 @@ class ProductSchema extends Schema {
       table.increments()
       table.string('category',100).notNullable();
       table.string('description',500).notNullable();
-      table.integer('price').defaultTo(0).notNullable();
+      table.float('price').defaultTo(0).notNullable();
       table.string('image').nullable();
       table.timestamps()
     })
